+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Stop official support of Neovim 0.6.
 - Use Lua API to create autocommands. Stop exporting functions only related to autocommands.
+- Use Lua API to create default highlight groups.
 
 # mini.base16